+++ src/main/java/com/blackrook/small/annotation/component/AfterConstruction.java
@@ -26,6 +26,4 @@ import com.blackrook.small.annotation.filter.FilterEntry;
  */
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface AfterConstruction
-{
-}
+public @interface AfterConstruction {}