+++ src/main/java/com/blackrook/small/annotation/component/AfterInitialize.java
@@ -26,6 +26,4 @@ import com.blackrook.small.annotation.filter.FilterEntry;
  */
 @Target({ElementType.METHOD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface AfterInitialize
-{
-}
+public @interface AfterInitialize {}